+++ BomberMan/src/main/java/it/bomberman/entity/creature/Creature.java
@@ -5,12 +5,12 @@ public abstract class Creature extends Entity {
 	public static final int DEFAULT_CREATURE_WIDTH = 228;
 	public static final int DEFAULT_CREATURE_HEIGHT =228;
 	public static final int DEFAULT_HEALTH = 2;
-	public static final float DEFAULT_SPEED = 8.0f;
+	public static final int DEFAULT_SPEED = 8;
 	protected int health;
-	protected float speed;
-	protected float xMove, yMove;
+	protected int speed;
+	protected int xMove, yMove;
 
-	public Creature(float x, float y, int width, int height) {
+	public Creature(int x, int y, int width, int height) {
 		super(x, y, width, height);
 		health = DEFAULT_HEALTH;
 		speed = DEFAULT_SPEED;
@@ -24,7 +24,6 @@ public abstract class Creature extends Entity {
 	}
 
 	// GETTER SETTER
-		
 	public int getHealth() {
 		return health;
 	}
@@ -37,23 +36,23 @@ public abstract class Creature extends Entity {
 		return speed;
 	}
 
-	public void setSpeed(float speed) {
+	public void setSpeed(int speed) {
 		this.speed = speed;
 	}
 
-	public float getxMove() {
+	public int getxMove() {
 		return xMove;
 	}
 
-	public void setxMove(float xMove) {
+	public void setxMove(int xMove) {
 		this.xMove = xMove;
 	}
 
-	public float getyMove() {
+	public int getyMove() {
 		return yMove;
 	}
 
-	public void setyMove(float yMove) {
+	public void setyMove(int yMove) {
 		this.yMove = yMove;
 	}
 }
