+++ BomberMan/src/main/java/it/bomberman/entity/creature/Entity.java
@@ -3,29 +3,29 @@ package it.bomberman.entity.creature;
 import java.awt.Graphics;
 
 public abstract class Entity {
-	protected float x, y;
+	protected int  x, y;
 	protected int width, height;
 
-	public Entity(float x, float y, int width, int height) {
+	public Entity(int x, int y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.height = height;
 		this.width = width;
 	}
 
-	public float getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
